nostrdb.c: use rcur instead of cursor for text search funcs.
This shows several things:
1. The rcur functions map naturally onto usage (e.g. returning NULL).
2. You don't need to worry about testing every return, just test if it
   was all OK at the end.

As a general rule:
1. Avoid `inline` inside .c files, since compiler will do that.
2. bool is clearer than int if it's really boolean.

+++ src/nostrdb.h
@@ -294,7 +294,7 @@ struct ndb_stat {
 // see `ndb_make_text_search_key` for how the packed version is constructed
 struct ndb_text_search_key
 {
-	int str_len;
+	size_t str_len;
 	const char *str;
 	uint64_t timestamp;
 	uint64_t note_id;
